SFAG seq 类型保持和正向一致



# message auto-generated for no-merge-commit merge:
!3762 merge sfag into master

SFAG seq 类型保持和正向一致

Created-by: dai-yuwen
Commit-by: dai-yuwen
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->

## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #000-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-transformer-dev!3762

+++ attention/sparse_flash_attention_grad/basic_modules/vec_op.h
@@ -235,11 +235,11 @@ __aicore__ inline void VecOp<SFAGT>::InitParams(const TILING_CLASS *__restrict o
             int64_t seqS1Len = 0;
             int64_t seqS2Len = 0;
             if (unlikely(i == 0)) {
-                seqS1Len = ((__gm__ int64_t *)actual_seq_qlen)[i];
-                seqS2Len = ((__gm__ int64_t *)actual_seq_kvlen)[i];
+                seqS1Len = ((__gm__ int32_t *)actual_seq_qlen)[i];
+                seqS2Len = ((__gm__ int32_t *)actual_seq_kvlen)[i];
             } else {
-                seqS1Len = ((__gm__ int64_t *)actual_seq_qlen)[i] - ((__gm__ int64_t *)actual_seq_qlen)[i - 1];
-                seqS2Len = ((__gm__ int64_t *)actual_seq_kvlen)[i] - ((__gm__ int64_t *)actual_seq_kvlen)[i - 1];
+                seqS1Len = ((__gm__ int32_t *)actual_seq_qlen)[i] - ((__gm__ int32_t *)actual_seq_qlen)[i - 1];
+                seqS2Len = ((__gm__ int32_t *)actual_seq_kvlen)[i] - ((__gm__ int32_t *)actual_seq_kvlen)[i - 1];
             }
             dimTq += (int64_t)seqS1Len;
             dimTkv += (int64_t)seqS2Len;
